+++ src/extension.ts
@@ -2,18 +2,8 @@ import * as vscode from 'vscode'
 import commandMap from './command-map'
 import {activateStatusBar} from './status-bar'
 import {activateShiftInterval, deactivateShiftInterval} from './shift-interval'
-import {
-  activateColorThemes,
-  shiftColorTheme,
-  favoriteColorTheme,
-  ignoreColorTheme,
-} from './color-themes'
-import {
-  activateFontFamilies,
-  shiftFontFamily,
-  favoriteFontFamily,
-  ignoreFontFamily,
-} from './font-families'
+import {activateColorThemes, shiftColorTheme} from './color-themes'
+import {activateFontFamilies, shiftFontFamily} from './font-families'
 
 export function activate(context: vscode.ExtensionContext): void {
   activateStatusBar(context)
@@ -22,30 +12,10 @@ export function activate(context: vscode.ExtensionContext): void {
   activateFontFamilies(context)
 
   context.subscriptions.push(
-    vscode.commands.registerCommand(commandMap.SHIFT_BOTH, async () => {
+    vscode.commands.registerCommand(commandMap.SHIFT, async () => {
       await shiftFontFamily()
       await shiftColorTheme()
       // FEATURE REQUEST: Reset shift interval
-    }),
-  )
-
-  context.subscriptions.push(
-    vscode.commands.registerCommand(commandMap.FAVORITE_BOTH, async () => {
-      const fontFamily = await favoriteFontFamily()
-      const colorTheme = await favoriteColorTheme()
-      vscode.window.showInformationMessage(
-        `Added "${colorTheme}" and "${fontFamily}" to favorites`,
-      )
-    }),
-  )
-
-  context.subscriptions.push(
-    vscode.commands.registerCommand(commandMap.IGNORE_BOTH, async () => {
-      const fontFamily = await ignoreFontFamily()
-      const colorTheme = await ignoreColorTheme()
-      vscode.window.showInformationMessage(
-        `Ignored "${colorTheme}" and "${fontFamily}"`,
-      )
     }),
   )