+++ src/status-bar.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode'
 import commandMap from './command-map'
-import {getColorTheme} from './color-themes'
-import {getFontFamily} from './font-families'
+import {getColorTheme, hasFavoritedColorTheme} from './color-themes'
+import {getFontFamily, hasFavoritedFontFamily} from './font-families'
 
 const STATUS_BAR_DISPLAY_TEXT = '$(color-mode) shifty'
 const STATUS_BAR_PRIORITY = 0
@@ -11,7 +11,11 @@ let statusBar: vscode.StatusBarItem
 export function activateStatusBar(context: vscode.ExtensionContext): void {
   context.subscriptions.push(
     vscode.commands.registerCommand(commandMap.SHOW_STATUS, () => {
+      const fontFamily = getFontFamily()
+      const colorTheme = getColorTheme()
+
       const actionTextMap = {
+        UNFAVORITE: 'Unfavorite',
         FAVORITE: 'Favorite',
         IGNORE: 'Ignore',
         SHIFT: 'Shift',
@@ -19,8 +23,10 @@ export function activateStatusBar(context: vscode.ExtensionContext): void {
 
       vscode.window
         .showInformationMessage(
-          `Using "${getFontFamily()}" font family`,
-          actionTextMap.FAVORITE,
+          `Using "${fontFamily}" font family`,
+          hasFavoritedFontFamily(fontFamily)
+            ? actionTextMap.UNFAVORITE
+            : actionTextMap.FAVORITE,
           actionTextMap.IGNORE,
           actionTextMap.SHIFT,
         )
@@ -29,7 +35,10 @@ export function activateStatusBar(context: vscode.ExtensionContext): void {
             vscode.commands.executeCommand(
               // eslint-disable-next-line no-use-extend-native/no-use-extend-native
               {
-                [actionTextMap.FAVORITE]: commandMap.FAVORITE_FONT_FAMILY,
+                [actionTextMap.FAVORITE]:
+                  commandMap.TOGGLE_FAVORITE_FONT_FAMILY,
+                [actionTextMap.UNFAVORITE]:
+                  commandMap.TOGGLE_FAVORITE_FONT_FAMILY,
                 [actionTextMap.IGNORE]: commandMap.IGNORE_FONT_FAMILY,
                 [actionTextMap.SHIFT]: commandMap.SHIFT_FONT_FAMILY,
               }[action],
@@ -39,8 +48,10 @@ export function activateStatusBar(context: vscode.ExtensionContext): void {
 
       vscode.window
         .showInformationMessage(
-          `Using "${getColorTheme()}" color theme`,
-          actionTextMap.FAVORITE,
+          `Using "${colorTheme}" color theme`,
+          hasFavoritedColorTheme(colorTheme)
+            ? actionTextMap.UNFAVORITE
+            : actionTextMap.FAVORITE,
           actionTextMap.IGNORE,
           actionTextMap.SHIFT,
         )
@@ -49,7 +60,10 @@ export function activateStatusBar(context: vscode.ExtensionContext): void {
             vscode.commands.executeCommand(
               // eslint-disable-next-line no-use-extend-native/no-use-extend-native
               {
-                [actionTextMap.FAVORITE]: commandMap.FAVORITE_COLOR_THEME,
+                [actionTextMap.FAVORITE]:
+                  commandMap.TOGGLE_FAVORITE_COLOR_THEME,
+                [actionTextMap.UNFAVORITE]:
+                  commandMap.TOGGLE_FAVORITE_COLOR_THEME,
                 [actionTextMap.IGNORE]: commandMap.IGNORE_COLOR_THEME,
                 [actionTextMap.SHIFT]: commandMap.SHIFT_COLOR_THEME,
               }[action],